move v2 into server
