--- serverv2/core/transaction/codec.go
@@ -1,5 +1,0 @@
-package transaction
-
-type Codec[T Tx] interface {
-	Decode([]byte) (T, error)
-}
